requirements.txt
Django==1.5
PIL==1.1.7
django-social-auth==0.7.22
sorl-thumbnail==11.12
